Fix the S3 backend so file uploads to GovCloud regions work.
Move the code that overwrites the file upload POST url into the
AmazonS3Backend class where it belongs because it's S3 specific. Modify
the code to reuse the provided hostname instead of hard-coding
s3.amazonaws.com which does not work for GovCloud regions. Add a new
rule to the URL interpolation whitelist to allow sub-region S3 URLs.