Fix a bug caused by cancelling the file selection dialog.
The FileInputDirective was not correctly handling the case where the
user opened the file dialog and then closed it by clicking the Cancel
button without selecting a file. Consequently the UploadFormController
was not handling that case either. This lead to a pretty harmless JS
error in the browser console when encountered.