Bind backend object in the DropzoneDirective.
The backend object created in BucketController is binded to the
Dropzone directive so it may be used to determine which backend
is being used (currently there is only one which is 's3').

This information is needed by the directive so the form data may
be modified to allow successfully making POSTs when uploading
files to S3.