Add function to AmazonS3Backend for setting form data.
When dropzone attempts to upload a file to S3, the form data needs
to be modified to include a 'key' parameter.

In this function, the 'key' parameter is set and the filepath takes into
account which folder the file is uploaded to and/or if the uploaded file
is located within a folder/sub-folder.