Remove folders that aren't deleted when "Show deleted files" is checked
I put the code that filters out the non-deleted folders in
loadContents() because it relies on getContents() to have already been
run and completed to produce accurate results. If getContents() hasn't
been run, this.folders could have stale contents or no contents.

I could see an argument for putting this into the
getDeletedContents method because the method name implies it
returns only deleted content, but it returns all folders regardless
of whether or not they are deleted which is misleading.

If performance isn't an issue, this logic could be moved to
getDeletedContents() if getDeletedContents() calls s3.getObjectsV2()
itself. In this case getDeletedContents() would no longer rely on
loadContents(), but the two separate calls to s3.getObjectsV2() could
return different results if a change to the contents of the current path
happens in between. Maybe that's not a big deal? (It occurs to me as I
write this that this problem could also occur if a change occurs between
s3.getObjectsV2() and s3.listObjectVersions())

In the end, I decided that having the method produce consistent
results was worth the trade off of requiring additonal post-processing.