Only display the latest version of a file when viewing deleted files
Prior to this change, if "Show deleted files" was checked then any file
that had a deleteMarker in its history would be displayed even if the
most recent version of the file did not have a delete marker. This
scenario happens when a file is deleted and then a new file with the
same name is created in the same directory as the original file.

Checking the isLatest flag ensures that the file will only be displayed
if the most recent version has a delete marker.