Change deleted file names to spans from links
Deleted files cannot be downloaded. Originally, deleted file names were
displayed as anchors blank href attributes. This produced odd behavior
in Django CMS (not sure if it actually affected live sites or not, but
disabling the CMS toolbar did not resolve the issue). The behavior
observed was the link was clickable and it downloaded the current page's
source code (as if the user had right-clicked and selected View Source).

While not a big deal even if this reached the live site, it was worth
changing to the deleted file names to be spans since they don't have
download links anyway.