Fix updating the working state.
According to https://stackoverflow.com/a/14049482, a new scope is
created when elements are created with ng-if. This means that two way
binding does not update outer scope attributes from the innerscope. This
is because the inner scope creates its own property for primitives which
hides the parent scope property. We work around this by passing a
function instead of a property.