Add backend function to upload file part.
A new function is added to AmazonS3Backend to upload a file
using AWS ManagedUpload. ManagedUpload handles chunking the
file and uploading each part.

This function expects the file and dropzone objects as parameters.
As parts are uploaded, the dropzone progress bar is updated
to reflect upload progress.

The function is named 'uploadPart' to remain universal for
various backends and not just AWS.