Remove function updateFormData from backend.
The updateFormData function was previously used to update the
formData by adding the 'key' field needed to make a valid POST to
AWS.

This function is no longer called since file uploads now go through
the backend uploadPart function which uses AWS ManagedUpload to
send the files.

In addition, the Dropzone 'sending' event handler no longer needs
to be defined since Dropzone is not used for sending files, and that
event handler never gets called.