Remove Dropzone 'sending' event handler.
This event handler no longer needs to be defined since Dropzone is
not used for sending files, and this handler never gets called.

In addition, the code which enable prompts when user attempts to
navigate away from the page during upload has been moved out of the
event handler and placed in the overridden Dropzone 'uploadFiles'
function.