Allow aws uploads to resume an existing upload
This will allow aws multipart uploads to resume if aws has a multipart
upload with a matching key.  For uploads with a matching key, the upload
id and info for completed parts are copied to the new upload.  If there
is no matching key for existing uploads, a new multipart upload is created.