Update file status and call processQueue in uploadFiles.
The Dropzone uploadFiles function needs to set the file status
and call processQueue after uploading a file (or attempting to).

The default Dropzone 'uploadFiles' function handles this in the
'_finished' function, but since we are overridding it, we need to
add those two lines into our function.

After uploading a file, we emit the 'complete' event and in the
'complete' event handler we check if we finished adding and uploading
all files. If we did, we then emit 'queuecomplete'.