Add component for use with sts credentials
This adds an STS session component to s3commander.  This component
is similar to the bucket component, but will take in an api url and
optional header.  This url must return the AWS credentials for an
sts session, including the session token and duration.  The component
itself will render a bucket with the temporary credentials.