NOTFORMASTER Test calculating SHA1 of a file before uploading.
In this commit, s3commander is modified to read an entire file
and compute the SHA1 hash before the file is uploaded. These changes
are made to test the effect of calculating the SHA1 hash of a file
in the frontend prior to uploading the file.

The changes in this commit support large files, as files are chunked,
and a rolling hash is calculated for each chunk. Once all chunks
have been read, the final SHA1 hash of the file is computed.

Debugging code exists in this commit which print out the calculated
hash of the file along with the time it took to read each chunk and
calculate the hash of each chunk.